Revert "Feature/typescript integration"
