+++ db/schema/seriesSchema.ts
@@ -1,0 +1,13 @@
+const mongoose = require('mongoose');
+
+export const seriesSchema = new mongoose.Schema({
+    seriesName: {
+        type: String
+    },
+    year: {
+        type: Number
+    },
+    rating: {
+        type: String
+    }
+});
