+++ graphql.config.js
@@ -1,6 +1,8 @@
-export const projects = {
-  app: {
-    schema: ["./data/schema.graphql"],
-    documents: ["**/*.{graphql,js,ts,jsx,tsx}"],
-  },
-};
+module.exports = {
+    projects: {
+        app: {
+            schema: ["./data/schema.graphql"],
+            documents: ["**/*.{graphql,js,ts,jsx,tsx}"],
+        }
+    }
+}
