+++ package.json
@@ -4,19 +4,16 @@
   "description": "graphql server project",
   "main": "./src/index.js",
   "scripts": {
-    "start": "tsc && nodemon ./dist/index.js --exec babel-node -e js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon ./src/index.js --exec babel-node -e js"
   },
   "author": "akrsh24",
   "license": "ISC",
   "devDependencies": {
-    "@types/express": "^4.17.13",
     "@types/node": "^15.0.3",
     "babel-cli": "^6.26.0",
     "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-0": "^6.24.1",
-    "tslint": "^5.20.1",
-    "typescript": "^3.9.10"
+    "babel-preset-stage-0": "^6.24.1"
   },
   "dependencies": {
     "apollo-server-express": "^2.24.0",
