--- src/data/index.ts
@@ -1,2 +1,0 @@
-export * from "./resolvers.graphql";
-export * from "./schema.graphql";
