--- src/db/index.ts
@@ -1,2 +1,0 @@
-export * from "./schema";
-export * from "./dbConnector";
