--- src/db/schema/friendSchema.ts
@@ -1,25 +1,0 @@
-const mongoose = require("mongoose");
-
-export const friendSchema = new mongoose.Schema({
-  firstName: {
-    type: String,
-  },
-  lastName: {
-    type: String,
-  },
-  gender: {
-    type: String,
-  },
-  age: {
-    type: Number,
-  },
-  language: {
-    type: String,
-  },
-  email: {
-    type: String,
-  },
-  contacts: {
-    type: Array,
-  },
-});
