--- src/db/schema/index.ts
@@ -1,2 +1,0 @@
-export * from "./friendSchema";
-export * from "./seriesSchema";
