--- src/db/schema/seriesSchema.ts
@@ -1,13 +1,0 @@
-const mongoose = require("mongoose");
-
-export const seriesSchema = new mongoose.Schema({
-  seriesName: {
-    type: String,
-  },
-  year: {
-    type: Number,
-  },
-  rating: {
-    type: String,
-  },
-});
