--- src/model/index.ts
@@ -1,2 +1,0 @@
-export * from "./Friend";
-export * from "./Series";
